Changed LoginImpl to not get from cache, since we always want to know if the login is correct and if so, store the credentials.
SettingsViewModel uses Helper method to send the logs

If login fails, added a way to send the logs to know what is going on.

Fixed ThrownIfEmpty: Added condition where the value could be an empty array of T (Empty Array of T is not the same as Observable.Empty<T>)

ReactivePresenter: Fixed issue when there is no empty or thrown exception, added a timeout.